added validation for actor + good practice by implementing service for actor and cleaning up the actor controller

+++ src/main/java/com/example/sakila/dto/output/FilmOutput.java
@@ -17,7 +17,7 @@ public class FilmOutput {
     private String description;
     private Year releaseYear;
     private Language language;
-    private Byte movieLength;
+    private Short movieLength;
     private String rating;
     private List<PartialActorOutput> cast;
